build: move books json generation to gradle task
