+++ buildSrc/src/main/kotlin/br/ufpe/liber/tasks/JsonExtensions.kt
@@ -1,0 +1,15 @@
+package br.ufpe.liber.tasks
+
+import kotlinx.serialization.json.JsonElement
+import kotlinx.serialization.json.JsonNull
+import kotlinx.serialization.json.JsonObject
+import kotlinx.serialization.json.JsonPrimitive
+
+fun String?.safeToJson(): JsonPrimitive = if (this == null) {
+    JsonNull
+} else {
+    JsonPrimitive(this)
+}
+fun String.toJson(): JsonPrimitive = JsonPrimitive(this)
+fun Number.toJson(): JsonPrimitive = JsonPrimitive(this)
+fun Map<String, JsonElement>.toJson() = JsonObject(this)
