+++ gradle.properties
@@ -2,7 +2,6 @@ micronautVersion=4.2.2
 kotlinVersion=1.9.22
 jteVersion=3.1.6
 luceneVersion=9.9.0
-exposedVersion=0.44.1
 flexmarkVersion=0.64.8
 kotestVersion = 5.8.0
